Position the re-use sub-menu under its button
When the reference to re-use has a lot of content the button element
of the sub-menu expands over the whole height of the result widget.

The expanded sub-menu will always be positioned at the end of the
button element, that's why it was always added to the end of the
result widget and not the visible button.

Makeing sure that the element of the button is not larger than the
visible outlines fixes that problem.

Bug: T375053
Change-Id: I3c16aeae37c3774808544b03c7e6e52762e6d145