Replace broken error when running out of backlink labels
This is the error message that appears when:
1. You replaced $2 with $3 in the message
   MediaWiki:cite_references_link_many_format.
2. A ref is used so often that we run out of labels in
   MediaWiki:cite_references_link_many_format_backlink_labels.

As of now, the error message has a series of major issues:

1. The error replaces the label. And because the label ends as
a parameter $3 in the middle of a wiki link [[…|$3]] the error
completely breaks the rendering of that link.

2. The same error renders over and over again for every backlink,
starting from the backlink where the labels run out.

3. From a technical perspective this is a very unexpected way for
the code to render an error, in addition to the many other ways
errors can render. Notably: Errors can either show up in the text,
replacing the footnot marker, or in the reference list. This is
using neither of these mechanisms.

After this patch:
1. Fall back to the numeric label.
   Render everything in a valid, usable way.
2. Show the error only once.
3. Use an existing error reporting mechanism.

In case you wonder about the ->plain() call. This doesn't mean that
the message was plain text. It's rendered as full-featured wikitext
later via Parser::recursiveTagParse(). This doesn't change with this
patch.

Also note that this mechanism is entirely unused in the Parsoid
code. This is by design.

Bug: T238061
Bug: T358770
Change-Id: Id893f44d8433fd013c4951077a570453fd177887