Switch to explicit numbering for Parsoid footnote markers
This patch removes all built-in CSS styling of Parsoid-generated
footnote markers, and reveals the explicit text instead.  The text is
generated using the same code as the legacy parser and should already
be "more correct" than what Parsoid currently renders, so any
behavior changes are expected to be an improvement.

The transitional rule can be removed once all on-wiki styling of
`mw-Ref` is disabled.

Bug: T382310
Change-Id: I39e6852cb600c57d1eec4dbb29dd02607c4248ca