Remove cite-tracking-category-cite-diffing-error
Using a category to suppress errors from visualdiff is brittle because
categories can be suppressed by individual wikis. In any case, the
information is already present on the Parsoid side (assuming the error
is emitted there), so we can suppress them that way.

We keep the entries in the localization file for now until the parser
cache has had a chance to expire existing entries.

Bug: T383141
Change-Id: If71d6d6b2b30c9031a35a60a6cb92ddf086e5234