Use configurable Alphabet to generate backlink labels
We're adding the underlying mechanics for auto generated backlink
labels here.  By default the auto generation uses the content
language for the Alphabet that will be retrieved from CLDR.

The Alphabet can be configured using $wgCiteDefaultBacklinkAlphabet.

When either CLDR or the config is missing it will fallback to `a-z`.

Currently hidden behind the $wgCiteUseNumericBacklinkLabels a
feature flag that would also replace the numeric auto generation
default when set to `false`;

Bug: T383216
Change-Id: Idd4dd62bb7e3317851fb4a967ab784fdde1956b4