References: Fix nested list-defined refs inheriting reflist's group name in Parsoid
Nested list defined references currently inherit the enclosing
reference list's group name if the "group" attribute is missing. This
causes blank references to be inserted into the current list, and
prevents the nested references from appearing in their proper reference
lists. The fix for this is to make ref tags' group names default to an
empty string (the default reference list) instead of the enclosing
reference list's group name, when inside of a "ref" tag. Which is how
the old parser's behaves.

In addition, adding an "inRefContent" method to ReferencesData to track
how deeply nested in ref tags the parser is. When nested references
occurs in reference lists, if the group name of the ref doesn't match
the group name on the reference list, an erroneous
cite_error_references_group_mismatch error is raised.
The error is raised near where I added the
"!$referencesData->inRefContent() &&" line.
That is, the following throws this error but shouldn't (granted, it
doesn't seem to be a visible error in most cases):
Text{{efn|name=note}}
{{notelist|refs=
  {{efn|name=note|Note<ref group="" name="ref">Ref</ref>}}
}}
{{reflist}}
The "ref" reference has a group name of "" while the reference list has
"lower-alpha" (set by the notelist and efn templates).
I'm not sure if this error was really present before, seeing as any
situation where this error would be raised would be a situation where
this very bug (T349981) would also occur and make the output
meaningless anyway. However, testing with the php "die" function and
seeing if the visual editor never loads on my test wiki, I can confirm
that these errors would have been raised, so, that's why I'm including
this fix as well. I don't want to be one to "fix one bug and introduce
another".

Originally, I had planned to just exclude blank group name references
from this check until I realized that, then, the following wouldn't be
an error when it should:
Text<ref name="ref"/>
{{notelist|refs=<ref group="" name="ref">Ref</ref>}}

The reason I chose to use helper functions "incrementRefDepth" and
"decrementRefDepth" instead of just directly modifying the
$nestedRefsDepth field I added is because, for nested reference lists
(which the old parser supports for some reason), $nestedRefsDepth needs
to be changed to an array (one item per nested reference list) so that
group checking isn't disabled for nested reference lists.
"$referencesData->incrementRefDepth();" seems more readable and
understandable than "$referencesData->nestedRefsDepth[0]++;" too, I
feel. I also find it easier to modify.

Finally, the reason I chose to add the $nestedRefsDepth field instead
of using "pushEmbeddedContentFlag" and "popEmbeddedContentFlag" is
because the latter will not work for nested reference lists at all
(because "inEmbeddedContent" checks if the value is anywhere in the
array, not if the first value equals it, so there would be no way to
re-enable the group check for nested reflists), and I assume simple
arithmetic on $nestedRefsDepth would have much better performance than
repeated "array_shift" and "array_unshift".

[Amendments]
(1) Upon further testing, it became apparent that a function like
"inRefContent" is required because "ref" tags that are direct children
(and not further descendants) of a reference list inherit the reference
list group name if the "group" attribute is missing. The function is
needed to tell when the default group name should be an empty string
versus the reference list group name.

(2) This same function is also needed to get rid of an erroneous
cite_error_references_no_key error which otherwise occurs when
references without names and follows are defined in a references list.
These are allowed in nested references. Due to the how tangential this
is to this patch, I feel that the hassle of having to add a failing
test and fix it in a second patch instead of including the fix in this
patch would be far too much of a hassle for what is effectively a one-
line fix.
As was exposed by the discovery of this (and my own testing, that I had
to use PHP's "die" function to test the
cite_error_references_group_mismatch errors I mentioned), for some
reason, these errors are not displaying. I thought this was just an
issue with the visual editor/edit preview, but, apparently, it is not.
That's something to be investigated and fixed later.

Bug: T349981
Change-Id: I0d3254563bae6300cb52a0c8341c059b0c1c5ce8