Fix "Jump up to:" text being shown with Parsoid
The reason for this issue is that Parsoid doesn't output a plain
text up arrow, but renders it as a ::before element via CSS. Our
little JavaScript module here doesn't only assume it would find that
text node, it relies on it.

Note this patch is really only a quick fix for the display issue.
The actual functionality of the highlighting module is still broken
in Parsoid.

Bug: T384204
Change-Id: I4b15adfe75e053df6620382fac4bf25fc451e1ae