Do not display errors on named references differing only by data-mw
There are a number of places where it is not enough to drop the
data-parsoid and about attributes to check for equality of references;
in case of templates and nested templates in particular, it can happen
that the typeof and data-mw attributes get dropped in some contexts and
not others, leading to confusing error messages.

Consider for instance the following wikitext:

{{1x|1=<ref name="ref">{{1x|<span>txt</span>}}</ref>}}
<ref name="ref">{{1x|<span>txt</span>}}</ref>

Both references contain the same content; but since the first one is
wrapped in a template, the span does not contain the nested template
information (whereas the second one does), leading to an error message.

The comparison content for two references intervenes in two places:
if the content is different, we display an error message, and we add the
HTML content of the inconsistent reference to the reference
datamw->body->html. If we do consider that two references are identical
even if their data-mw is not, we risk losing information that would have
otherwise been added said html data-mw attribute.

Hence, in this patch, we do a two-step comparison: we check if two
citations are equal "up-to data-mw and typeof" to decide whether to
display an error message, but we keep storing the reference's HTML if it
differs on the data-mw status from the "source of truth" HTML.

Note that this patch is not solving the issue entirely:

{{1x|1=<ref name="ref">{{1x|txt}}</ref>}}
<ref name="ref">{{1x|txt}}</ref>

will still trigger an error, because a wrapping span is generated in one
cas and not the other, which still triggers a diff. But, this case
(which feels like it's on the fringe of the border of the corner case)
can be worked around by wrapping the content of the reference in a span
on the wikitext level.

Bug: T380152
Change-Id: Iceee0c34b5fc72d069c1c4ba4aa779ef21c5f3c8