Add Codex UI for the CommunitConfiguration backlink marker settings
Here we add a first Codex version to read and set the backlink marker
setting using the CommunityConfiguration provider. The interaction
is pretty plain and simple. Further elements to the UI and workflow
will be added in follow ups.

The whole UI is hidden behind the feature flag. The setting will
not yet be used for generating the backlinks.

Only admins can save changes.

Bug: T378807
Change-Id: I012d21d2f2cb20055a7721259ee155d70979ee39