Do not create a linkback for references that are children of themselves
In some corner cases of the <ref name="a"><ref name="a">content</ref>
pattern, creating a linkback and removing the corresponding fragment
induces keeping a reference to a deleted node in the "$ref->nodes" list,
which throws an exception in PHP 8.1 (although that is not the case in
PHP 8.3) down the line when it is accessed. We fix this by not creating
the linkback in the child node in this case (it would not make sense
anyway) and adding a cite_error_included_ref instead.

The identified pattern that triggers this issue is

{{1x|
<ref name="ref1"/>
<ref name="ref1">some content
* <ref name="ref1">ref1</ref>
</ref>
}}

although other variations may exist in the wild (this issue
triggered on at least two enwiki pages.)

Bug: T382546
Change-Id: Icedd6fdf718183ad5bb873a9f92421f61aa49217