Add basic support for indirect sub-references by name
There are a few open questions left, e.g.:

Shouldn't the anchors of the sub-ref repeat the name of the parent
reference? At the moment the anchors are build as if the sub-ref is
unnamed.

Still I believe this is a good step forward and can already be
merged as it is.

Bug: T385570
Change-Id: Ib0c7f66c76f6fe63053fbcd7a8de62a0f919e01f