Try to clarify the meaning of "key" accross the codebase
There is an absolute ton of different meanings of the word "key" in
this codebase. Probably 5 or 6 different meanings. There is even
very, very old code left where "key" was used for the name of a
<ref>.

This patch doesn't manage to clean up everything. But I had to stop
somewhere to keep this in a state where it can still bve reviewed.
To be continued in later patches.

Bug: T386452
Change-Id: If81deb567895a1dbdd72e833162ea10cf0408be5