Use the name instead of the internal array key where possible
Turns out these array keys are meaningless. For named <ref> it's just
the name and we can use $ref->name instead. For unnamed <ref> it's
a meaningless number that – as it turns out – is not used for
anything anywhere. There is one exception in the rollback code. But
that code doesn't care what the array key is. It could be anything.

TL;DR: In all situations where that $key is used the $ref->name is
guaranteed to be set.

Bug: T385570
Change-Id: Iad40e13f910afdfc95d1cdbae73064d2f0111c79