Drop redundant backlink id property from Parsoid code
Turns out this property is completely redundant. For unnamed <ref>
it's identical to the other property. For named <ref> it holds
a copy of the first value with a hard-coded "-0". But that can
easily be generated when needed. And it's only needed in a single
place anyway.

This is heavily covered with tests. This is key. As long as all
tests still succeed we are pretty much free to do whatever we want,
as long as it makes the code easier to understand and easier to
maintain.

Bug: T386452
Change-Id: I04bd2260329e84b8c8d59a326cb17609a4fbdc21