Index by globalId and avoid slow loop in rollback code
It's still beneficial to index named refs by name, because this kind
of lookup is done in more than one place. But there is only a single
place where such a lookup is done for unnamed refs. That was a loop
before but can easily use the global numeric id instead.

Bug: T386452
Change-Id: Ie870acc46b77cd8e3a09c35ccd0eaff63a8a6bed