Drop not needed "assign" action from ReferenceStack
Look at the rollback code. The only thing the "assign" action does
is unsetting the text. But why? When there is no conflict in the two
text values it doesn't matter if two consecutive refs look like

 <ref name=a />
 <ref name=a>x</ref>

or

 <ref name=a>x</ref>
 <ref name=a>x</ref>

Both behave exactly the same. We don't need to distinguish between
them in the rollback–redo code. There is no code anywhere that needs
this information. As proven by the parser tests.

The unit tests are affected because they test private implementation
details. This is intentional. These tests are meant to be regression
(smoke?) tests.

Bug: T387991
Change-Id: I1af940d95227f716328712a82d6ac54dac264d34