Add error when combining details with follows
Note this message was intentionally designed to cover multiple
scenarios, see Id94e177.

You might wonder what the difference between name="0" and
details="0" is? The name "0" is invalid anyway and rejected by
another existing validation step. The same is not true for details.
We need to make sure the string really is empty (or not empty) at
this point.

Bug: T387003
Change-Id: I24e16f1942387ea99580f7c466098cdfa9b3a4be