Re-classify one "details" validator error as a warning
This introduces a 2 tier system in the existing Validator:
Fatals/errors are meant to invalidate a <ref>. It is not registered
and never rendered. Instead the user only sees the error. This is
the existing behavior. The biggest problem with this approach is
that it causes all kinds of other errors because of the missing ref.
That can be very confusing.

New behavior: In case there is no fatal/error but only a warning
the ref is registered and rendered as if it is valid, just with the
warnings appended to it.

For now this is touching only a single error message that was added
via I96d1ea6 and is currently not deplyed anywhere. However, the
idea is to do the same with a bunch more error messages, even older
ones.

Bug: T238061
Bug: T387003
Change-Id: Ib0683ff9577da953545a6ba54fb964bc63b307c3