Re-classify invalid dir="…" attribute validation as warnings
Instead of making the <ref> fail and not register now we only drop
the attribute. I wanted to fix this for a very, very long time.

This patch is intentionally only touching a single message. Kind
of as a proof-of-concept and to prepare the code for all the other
messages (see the following patches). Note this is a rare message
users typically don't see and don't care much about. It was not
properly implemented for a very long time, see I4ab0cb5. I believe
we are free to merge this any time without an announcment.

Bug: T238061
Change-Id: I00bf95b27052ec867aaee6feae867dd21697e7f9