Fix duplicate errors on sub-ref and main reference
This was just a mistake. It doesn't make much sense to clone the
errors and show the same errors on two refs. The reason we did not
run into this so far is that most errors are still hard fatals and
stop the rendering. But we started to change this and turn them into
softer warnings.

Bug: T387003
Change-Id: I0406fd8638d44a5ceabe2aabf01f03ed9f9ac49b