Drop now obsolete parameter in internal rollback stack
This is a direct follow up to Ie870acc. Since then we need only one
information: What the array key in the internal $this->refs array is.
It's either the ref name for named refs or the globalId for unnamed
refs. But the rollback code doesn't care if it's one or the other.
We can merge the two elements into one.

Bug: T386452
Change-Id: I9597d62b7c5ee9434ed2dbac62b64b5ea767df6e