Fix rollback logic to roll back main and sub-ref
Since I4f31f70 we have an action that creates 2 entries in the
internal $this->refs structure the same time. Both must also be
rolled back in one step. Previously the code tried to do this with
an extra entry in the internal rollback stack. But again there is no
code that would understand that it sometimes needs to process these
2 entries in one step.

I tried a few approaches:
* A new action. Unfortunately this alone doesn't help because we
  need the lookup information for both entries.
* Storing 2 lookup entries in the $call[2] slot (see older
  patchsets). This works but felt ugly.
* Just use the existing lookup where the subref points to its
  parent anyway.

Bug: T388131
Change-Id: I4a7118cd27b27c5120b1693a5b70b0ffc500a19e