Fix extra end of the line whitespace in parser tests
I constantly run into this whenever working on anything related to
parser test and my IDE removes seemingly meaningless whitespace from
the end of the line. Which makes this test fail. Instead we can
modify the test input (the HTML) a little bit to avoid this problem
entirely. The test still covers what it is supposed to cover.

Bug: T386792
Bug: T387014
Change-Id: I1da517d5f91f64467ad6d485fd15e002e93d0739