Rename ambiguous "index" and make it behave the same
Make the behavior the same in both parser implementations. Both
count the total number of references processed so far, and assign
ids starting from 1.

This continues what we started in T386452.

Note that no (parser) test changes. The behavior of the code really
does not change. The field is especially still unused. We might
change this in a later patch.

Bug: T387991
Change-Id: I72db9525f31626414473ead2310e5edf6ae24f80