Start using the shared AnchorFormatter in Parsoid
This gets rid of two properties in the RefGroupItem class.
Instead we use the same AnchorFormatter as in the legacy parser to
dynamically generate these anchors only when we really need them.
Note how much code can be removed from various places that was
generating (partial) anchors before.

The is no change in behavior, as proven by the tests. I'm expanding
one parser test only to cover the previously uncovered extra
escaping code. This is only to document the existing behavior.

Bug: T387991
Change-Id: Ida9e9a284dad98c7bfe71a5122b3ba1becaea90b