Remove not needed extra rollback stack entry
This is only a first step.

Old issue: One <ref> with details creates 2 entries in the rollback
stack. But there is no code that can roll back these 2 entries. This
messes everything else in the stack up. The test demonstrates this.

New behavior: Create only 1 rollback entry for the subref. This means
that the parent ref will not be rolled back. Is this an issue? Maybe.
I found a minor edge case and documented it in a new parser test. We
will continue working on this. Still this patch is an improvement.

Bug: T388131
Change-Id: Ia7839aacae568086e434e806c7af2ec7b973aa0f