Higher priority for "no details in <references>" error message
This was originally implemented via I96d1ea6. I had to push the
warning down to not break the existing assumption that all <ref>
inside of <references> have a name. That assumption would have been
broken by the new warning. But this is not an issue any more since
I4bac62c.

Note this is not the absolute highest priority. There are still some
combinations that come before the details warning, e.g. a numeric
name.

The other change is a slight modification of the second message
specific to details, as suggested by PM and other devs. We are still
free to change this message as it's not deployed anywhere.

Bug: T387003
Bug: T388570
Change-Id: I8ee6bb0381db9a5a72be75d59a3484f04ebb454f