Add parameter name to "invalid argument" error message
I played around with a few options and the most trivial one that
covers almost all relevant situations is to just show the first bad
argument. While it's possible to have more than one this is rare and
makes the code very complex for not much benefit. The user can still
move forward and fix one error at a time.

Bug: T388497
Change-Id: If7ebfdff5378a8aee68a7e05050f5eb76700fc6b