Fix unused sub-ref parent sometimes getting a backlink
I added this code while working on T385570. While reviewing Ifbc4295
I found another related, broken edge-case.

 <ref name="a" details="x">x</ref>
vs.
 <ref details="x">x</ref>
should behave the same but currently doesn't. Both render almost
identical. The difference is a misplaced backlink in the main ref.
As shown in the tests.

Note this is unrelated to the discussion in T387193. This is really
only fixing a tiny mistake in the legacy parser.

Bug: T385570
Bug: T388131
Change-Id: Id639c4f6afca828d88f955849ab59ff02e3b4b1a