Make the validator return all errors, not only the first
At the moment this doesn't make much of a difference. The only
difference is if there is a (soft) warning first, followed by a
fatal error. The old code would only report the warning and
suppress all following fatals. The new code still reports the same
warning first, but with the $status->isOK() flag set to false.
This results in the <ref> not being rendered, as it should.

This is critical for the proposed change in I8ee6bb0.

Bug: T387002
Bug: T388570
Change-Id: I4bac62c827d52a94365eaabcd0e0765723509042