Move first argument validation step into the validator
This already allows us to start using it in the Parsoid code.

Again this is only a step in the process. For the moment the new
method is static. This is not ideal and only meant as an incremental
step forward. Ideally everything should be merged in a single
non-static validation step.

Bug: T387002
Bug: T387991
Change-Id: I889871838d3a1b0f95d0d8ebb1f139e82f7f0389