Move all (back)link generator logic into AnchorFormatter
Look just how much actual logic was living outside of the class,
scattered around the rest of the codebase. Now it's so much more
obvious what's actually going on. Including a weird special case
for the follow feature that doesn't conform to the rest.

Finally there is no mention of a "key" any more in the class.

I'm mainly doing this for two reasons:
* I hope this makes it possible to reuse the class in Parsoid.
* Even if not, now it's so much easier to understand how these links
  are generated and make them behave the same in both parsers.

Bug: T386452
Bug: T387991
Change-Id: I14788000e083907ee5c7c2b01a5ced5598728ab9