Rename ambiguous "target" to "noteId" in the Parsoid code
We do have "targets" in both directions, when jumping down to the
reference list and when jumping back up.

This is the last remaining field we have not renamed before as
part of T386452. I was not sure but got inspired by Ia86b1f6.

Also move the name field up in RefGroupItem, next to the group.
It had a quite confusing position before.

Bug: T386452
Change-Id: I98a6359023cb605a7473328576ccd74cf3bf68b8