Remove non-standard id attribute from incomplete follow
The code is like this since the original patch that introduced the
feature in 2010. It's unclear what the purpose of this id attribute
ever was. By definition there is nothing that links to this id.
The <p> is an incomplete follow that's not used on the current page
but on e.g. the previous page. When the pages are rendered together
the <p> is resolved and everything works just fine. This is
unaffected by this patch.

A manually crafted [[#…]] that links to this non-standard id wouldn't
make sense because it can only work in a single-page context but not
when the pages are rendered together. The id is different then.

My first idea was to just remove the special case and let the code
generate another id. But that would have no benefit.

We can use
https://global-search.toolforge.org/?q=cite_note-%5Ba-z%5D&regex=1&namespaces=102
to check if such links exist anywhere in a Wikisource project.
I cannot find any.

Bug: T387991
Change-Id: I1c40e8762d037394345df35d4c13f02085c52bbf