More consistent, stricter error handling with empty details=""
This is only relevant in edge-cases where a non-empty details="…"
would trigger an error anyway. These are:

1. details without main content and without a name is missing its
   parent. Not reporting an error in this situation means we would
   silently render this as a normal ref, even if it's marked as a
   sub-ref.
2. details cannot be used inside of <references>. It's effectively
   the same as any other invalid attribute in this context. They
   all trigger an error just because they are present, even if
   they are empty. details just uses it's own error message.
3. follow cannot be used together with details. Same as above.

In situations where a non-empty details="…" is allowed an empty
details="" is also allowed.

It's easy to relax one of the listed edge-cases later if we need
to. But it's much harder to make the validator more strict later,
after people already started using the new feature.

Bug: T386212
Bug: T387002
Change-Id: I1ad49e085d9b1d9967ec4ed0dc75d2bbd2e11319