Add tracking category for sub-references
Implement tracking category for pages that use the new details
attribute: <ref details="…">. The tracking category is set when
the attribute is used at least one time per article, even if the
attribute is empty. It's not set when the feature is disabled,
even if people try to use the attribute.

Not part of this patch: Parsoid should do the same. A separate
patch will follow.

Bug: T388762
Change-Id: Ifc2ed2aacbf5a082667c266a52ce92e47f823ee4