More consistent, isolated test setup in integration test
This is motivated by the work in T388762. The idea is to make the
different test cases behave more consistent. They should all use the
same methods to create Parser as well as Cite instances.

This also makes the test a little more isolated. E.g. there is no
actual config any more but only a local HashConfig.

Bug: T388762
Change-Id: If377ab4b80d1452ec58bcad23b48a14099bbcb5e