Disable unnamed sub-refs in legacy parser as well
This might (or might not) be temporary. In any case we want the two
parsers to behave identical to reduce the amount of confusion. Since
Ifbc4295 Parsoid just ignores the details attribute when there is
no name. There is no error message in Parsoid. This is fine in the
work-in-progress code in Parsoid but not here in the Validator
class.

Bug: T387193
Bug: T391472
Change-Id: I2125f251b293395df01fe87c493ae87e3d8f2c01