Fix double escaping on CommunityConfiguration Vue page
We found this issue when we reviewed T385411. The output is escaped
anyway, but calling .parse() escapes it twice, resulting in &#…;
encoded characters becoming visible.

I'm also using the oportunity and order the messages the same way
they appear in the UI. This makes the job of the translators on
translatewiki easier.

Bug: T384495
Bug: T385411
Change-Id: I02253e411de6724d3d2dfaf88122714c8fdefadc