Isolate conflicting content check
Not exactly a pure function since it has the side effect of caching
some content inside of the ref, but still beneficial for reducing
complexity in extractRefFromNode.

We ended dropping and rephrasing the previously used concepts of
"different HTML" and "different content". Both are quite ambiguous.
What kind of HTML? What kind of content?

What actually happens (slightly simplified):
* Different "content" is any difference after dropping data-mw
  attributes. This includes differences in e.g. class attributes.
* Different "HTML" is anything but different content. That can only
  happen when there is a difference in data-mw metadata, but the
  visible HTML is the same.

I go with "visible" in this patch. Other options are:
* Something with "different metadata"?
* Something with "different source" vs. "different rendering"?

Bug: T390960
Change-Id: I9d14c6e4b3e6d9938def298eb6fabf6020595371