Conflict detection is local to References class
Makes it clearer that conflicting HTML knowledge is only relevant to
one specific function, this isn't data that needs to get passed
around along with the ref.

The data structure is unfortunate, but my thinking is that slightly
increased complexity within this function is an improvement over
potentially distributed complexity, ie. having to wonder about
long-term effects of this cached variable.

Bug: T390960
Change-Id: I01c01bc72e0b4402128c295549b3a2cd0227f6ea