Move misplaced domToHtml out of the validation code
This is a direct follow up for what we found in I9d14c6e. It looks
like these calls can safely be moved down without making any
difference in any of the test scenarios. The main benefit is that
the extra (expensive!) domToHtml call is done much less often.

I have to say that I don't fully understand this code. But I trust
the tests.

Bug: T390960
Change-Id: Id79eab46b8020a20c86a81d15b7122d78f8831dc