Move Parser dependency into the ErrorReporter constructor
This actually moves the dependency a little bit *deeper* into the
class, away from the individual calls up into the constructor. I
would argue that this is fine. It makes a difference for quite a
lot of unrelated code in other classes that don't need to pass the
parser around any more. The ErrorReporter itself heavily depends on
the parser anyway. There is no way we could ever use this class in
Parsoid.

Change-Id: I31695c04301c7558d4adb50ec36cf785ae190024