Move extra error checking out of ReferenceListFormatter
A formatter should not actively add new errors. That's confusing.
It also doesn't make it easier to reuse this code.

This patch also contains a little bit of directly related cleanup in
the Parsoid code.

The idea is to make it easier to possibly move this code to the
validator where the two parsers can reuse it. This will happen in a
later patch.

Bug: T387991
Change-Id: Ic87e933b176a8d6012e8dc65989d8f4127853589