Consolidate duplication and use consistent attribute order
The two attributes "mainBody" and "isMainRefBodyWithDetails" used a
different order in the two code paths. I think it's not relevant
which order we use, but I had to make it consistent to be able to
merge the duplicate code paths.

It also looks like it doesn't matter if there is a $contentId, or
can't happen in the first place.

This is motivated by I138d930 and hopefully makes it easier to fix.

Bug: T390960
Change-Id: I0e64c2ec1815019a44cc584985dc5adf26f4351b