VE: Extract content set logic in toDomElements
Two parts in the method are about if the current node should get
the body content when writing the DOM.  Here we tried to rewrite
and extract these two.

We also made sure that nodesWithSameKey is always a list. When
iterating over the items we do not need to check if its truish.

Bug: T392577
Change-Id: I1ef401653d32d5c0974747b8e6f161291c87a493