Rename slightly confusing "isMain…WithDetails" attributes
There are two very similar attributes that are very easy to confuse.

Also turn them into shorter ints while we are here. Using true would
also be possible but would be stored as `true`.

We are still free to rename these attributes. Notice how only the
(not yet deployed) sub-ref tests are affected.

Bug: T389929
Change-Id: I4a057f04d166b80e1d5d694f151344b223f75cea