Fix cite_error_references_missing_key differences in Parsoid
This removes the last exclusion. Now the Validator is used 100% the
same in both parsers.

Note this change appears to remove an error message from a few
test cases. This is intentional. These test cases are all about a
follow="…" where the parent ref cannot be found. This is not an
error but an expected situation.

follow="…" is designed for Wikisource where books are digitized,
page by page. It's actually possible for books to contain footnotes
that are so long that they are split accross multiple pages. When
digitized the first page gets a normal <ref name="…"> with a normal
marker [1] in the text. The second page doesn't get a [1] but only
the continuation of the footnote. This is done via <ref follow="…">
in wikitext.

When the pages are rendered together the two ref are merged. But
it's also possible to view the pages individually. The fact that
each contains half of a footnote is not an error!

The old Cite code renders the second half of the ref as a <p>.
(Technically misplaced inside of the <ol>. But it works.) Parsoid
currently renders it as a <li>. That's not the same, but it's not
the goal of this patch to remove this discrepancy.

What this patch does is removing the error message that's currently
only in Parsoid, but never was in the old parser.

Bug: T393022
Bug: T398402
Change-Id: I01860775ecc301ba638fca2ebf8c28944a73131c