VE: Reusing a sub-ref in VE creates an independent copy
We currently don't want sub-refs to be merged to align with the
behavior in reader mode.  Therefore we're just adding a copy each
time the user reuses one.

For this we're adding a method to the ve.dm.MWReferenceModel that
can also be used in the Citoid inspector where the same needs to
happen.

Copy&Paste relies on a different method and can be fixed later.

Bug: T392724
Change-Id: Ibe63176883972bef671ef7da734544baf07e3a6a