VE: Make more use of makeListKey
I refactored that method a bit to make more use of it.  It's also
to prepare the copy of sub-refs when using C&P where we want to
generate a new listKey for the copy.

Also using `this` in the places where this method is used. They
are all static and the code looks a bit nicer with that.

Bug: T392724
Change-Id: I8c0af5b10d55880b65b2499a3ec77c645f1ed442