Replace many .html() calls with .append()
This is mostly about the readability of the code and should not
have any actual consequences. Many of these .html() calls have been
added in Id8684cc. I believe at least one was just a mistake.

This patch also renames one of the methods for more clarity. There
is a bunch of methods with the same name. This makes the code rather
confusing.

Please carefully test this. This code is not fully covered with
tests. I tried to test it as good as I could but might have missed
something.

Change-Id: I6d36466b803c748aaeb6214731b44d8087892245