Move "empty body in <references>" error check to the Validator
There is a single actual change in the test output. As far as I can
see the new behavior is more correct than it was before.

Note that the legacy parser only outputs one error message (the
first it can find) while Parsoid is able to output more than one.
The change we see is only one of these extra errors in a situation
that already had errors.

Bug: T393022
Change-Id: Ib7b0122619d890a5d7a230cdd23a17d4602d086a