Re-classify "invalid <ref> parameter" from fatal to warning
This is (almost) entirely harmless. Extra parameters in a <ref> tag
never do anything anyway. They cannot have any effect on anything.
There is not much reason to suppress a <ref> entirely just because it
had a mistake in a parameter name, e.g. when gruop="…" contains a
typo. Instead we can continue and render the <ref>, just with an
error message attached.

This patch moves the error message down into the reference list.
It's still rendered as before. Community customizations continue to
work as before.

The only effect worth mentioning is on wikis that currently suppress
error messages via CSS. On such wikis the <ref> will appear with no
visible error. Still this is an improvement compared to the current
situation where neither the <ref> nor the error message is visible.
That community decided to use a different process to fix such errors:
The error category is set and a some expert users go around and fix
such articles. This process is untouched by this patch. Fixing the
issue even becomes a little easier when the <ref> is easier to find.

This is relevant for sub-referencing because we expect users to play
around with the details="…" parameter and run into this message more
often.

Bug: T238061
Change-Id: I5f0a295089853055c98a879806cbc92da47f814c